+++ net/ATL.Servicio/ATL.Entidad/TallaPedido.cs
@@ -11,6 +11,6 @@ namespace ATL.Entidad
         public long idPedido { get; set; } //temporal
         public long codigoPedido { get; set; } // central
         public int numeroTalla { get; set; }
-        public int cantidad { get; set; }
+        public long cantidad { get; set; }
     }
 }
